+++ kraken/time_actions/common_time_functions.py
@@ -5,6 +5,8 @@ import kraken.invoke.command as runcommand
 import kraken.kubernetes.client as kubecli
 import re
 import sys
+import kraken.cerberus.setup as cerberus
+import yaml
 
 
 def pod_exec(pod_name, command, namespace):
@@ -134,3 +136,17 @@ def check_date_time(object_type, names):
             if counter < max_retries:
                 logging.info("Date in pod " + str(pod_name[0]) + " reset properly")
     return not_reset
+
+
+def run(scenarios_list, config, wait_duration):
+    for time_scenario_config in scenarios_list:
+        with open(time_scenario_config, "r") as f:
+            scenario_config = yaml.full_load(f)
+            for time_scenario in scenario_config["time_scenarios"]:
+                object_type, object_names = skew_time(time_scenario)
+                not_reset = check_date_time(object_type, object_names)
+                if len(not_reset) > 0:
+                    logging.info("Object times were not reset")
+                logging.info("Waiting for the specified duration: %s" % (wait_duration))
+                time.sleep(wait_duration)
+                cerberus.publish_kraken_status(config, not_reset)
